Lab05/lab5_2/kafka_lab05/pyproject.toml
[tool.poetry]
name = "kafka-lab05"
version = "0.1.0"
description = ""
authors = ["Tomás Santos <t.santos.f@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.0.2"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
